image_converter/pyproject.toml
[tool.poetry]
name = "image_converter"
version = "0.1.0"
description = ""
authors = ["megahomyak <g.megahomyak@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^10.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
